seccomp-ebpf: Add ability to read user memory
This uses helpers bpf_probe_read_user{,str}. When the loader of the
filter does not have CAP_SYS_PTRACE, attempting to read user memory
when current mm is non-dumpable results in -EPERM.

The check of CAP_SYS_PTRACE is against the init userns. This is
because the eBPF filters cannot be loaded from any other userns.
[Jinghao Jia]

Right now this is not sleepable, -EFAULT may happen for valid memory
addresses. Future work might be adding support to bpf_copy_from_user
via sleepable filters.

Use of memory data to implement policy is discouraged until there is
a solution for time-of-check to time-of-use.

Signed-off-by: YiFei Zhu <yifeifz2@illinois.edu>
Co-developed-by: Jinghao Jia <jinghao7@illinois.edu>
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>