bpf, seccomp: Add eBPF filter capabilities
This introduces the BPF_PROG_TYPE_SECCOMP bpf program type. It is meant
to be used for seccomp filters as an alternative to cBPF filters.

This commit only adds the new program type, i.e. the necessary
implementation to support eBPF filters. The program type currently
cannot use any maps/helpers. The support for maps/helpers would be
added in later patches.

The loading of eBPF filters requries CAP_BPF and CAP_PERFMON in the
init userns. The attachment of filters is consistent with the current
seccomp with cBPF filters, which requires the process to have either
CAP_SYS_ADMIN in its userns or the no_new_privs bit set.
[Jinghao Jia]

The capability requirement for loading also implies that the eBPF
filters cannot be loaded in any other userns, as processes would lose
their capabilities in the init userns when entering a new userns.

The eBPF code loading is separated from attachment of the filter, so
a privileged user can load the filter, and pass it back to an
unprivileged user who can attach it and use it at a later time.

In order to attach the filter itself, you need to supply a flag to the
seccomp syscall indicating that a eBPF filter is being attached, as
opposed to a cBPF one. Verification occurs at program load time,
so the user should only receive errors related to attachment.

The behavior of eBPF filters with bitmap cache is that they will
totally negate the bitmap acceleration. Static analysis of eBPF to
create a bitmap could be a potential future work. [YiFei Zhu]

rcu_read_lock is also held because maps like BPF_MAP_TYPE_LRU_HASH
requires them. For simplicity, this also affects cBPF filters.
[YiFei Zhu]

Signed-off-by: Sargun Dhillon <sargun@sargun.me>
Link: https://lists.linux-foundation.org/pipermail/containers/2018-February/038572.html
Co-developed-by: Jinghao Jia <jinghao7@illinois.edu>
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>
Co-developed-by: YiFei Zhu <yifeifz2@illinois.edu>
Signed-off-by: YiFei Zhu <yifeifz2@illinois.edu>