seccomp, ptrace: Add a mechanism to retrieve attached eBPF seccomp filters/maps
This extends the ptrace API to allow fetching eBPF seccomp filters
attached to programs. This is to enable checkpoint / restore cases.
The user will have to use the traditional PTRACE_SECCOMP_GET_FILTER
API call, and if they get an invalid medium type error they can switch
over to the eBPF variant of the API -- PTRACE_SECCOMP_GET_FILTER_EXTENDED.

It also adds the functionality for retrieving eBPF maps associated with
a certain filter and task. In this way checkpoint / restore could use
the map fd to dump maps (CRIU currently supports doing so for array and
hash maps.)

Signed-off-by: Sargun Dhillon <sargun@sargun.me>
Link: https://lists.linux-foundation.org/pipermail/containers/2018-February/038478.html
[YiFei: increase ptrace number to 0x4210]
Signed-off-by: YiFei Zhu <yifeifz2@illinois.edu>