Implement bpf_seccomp_close_fd in libbpf
This commit adds the new bpf_seccomp_close_fd() function in libbpf.
It is specifically for seccomp-eBPF filters that it closes all
fds (e.g. to maps, BTFs) except the first fd of the first program
in the struct bpf_object returned by bpf_prog_load(). The fd left
open should be used for seccomp filter installation, where it will
be closed automatically.